BUG/MEDIUM: api: exclude DELETE calls from content-type check (#74)
In 816bbfd98b (add content-type and redirect handling) some content-type handling
was added to validate that we were receiving a proper JSON response back.
However, in the case of DELETE calls there is an empty content-type returned.

This causes erroneous errors such as the following:

```
│ Error: received unexpected content-type () response:
```

This commit excludes DELETE calls from the content-type validation.