BUG/MINOR: api: ignore content-type check for http/204 (#76)
In 9cbaa52 (BUG/MEDIUM: api: exclude DELETE calls from content-type check) logic
was added to disregard checking if the content-type is application/json for the
DELETE method. While this fix did address the issue at hand it was not as
extensive as it should have been.

It was recently reported that PATCH methods are also having an issue. Instead of
adding each individual method it is better for us to check if the HTTP status
code is 204 (No Content) as the API is returning this when it is not expecting
to return any content.

This commit swaps the method check with one on the response code 204 and should
fix this issue broadly against all methods.